test: fix proxies so it stops breaking tests
Modified proxy to return the same jest mock functions on property
access. Otherwise we won't ever be able to do any assertion on it
if it returns a value dynamically every time.

Also, only load in jest mocks only in test env. This allows the mocks to
be shared between tests and Storybook. Otherwise, Storybook doesn't seem
to like importing jest and using jest mock functions