Make editable radius marker
I am going to add comments and clean it up later. This took bloody forever and required workaround after workaround.

At first I wanted to drag the circle anywhere to resize it. However the library doesn't support this, you can only drag to move, or it tries to move and resize at the same time. So I had to go with a second small "handle" which can be moved to resize.

However it's still not perfect, moving the circle resets the marker "angle" back to 0 degrees.