Complete draggable radius
Mapbox GL JS has a problem where clicks on overlapping layers fires events for all the layers. It is possible to check if the event has fired before, but not possible to choose the order. In this case, the circle was always firing before the marker/resize handle, when the opposite would be is important for UX.

As a workaround, I offset the resize handle so there is no overlap, but this makes the code rather verbose. I don't see another way around this but I would love to hear any possibilities.