Cleanup - Complete adjustable radius
The radius has a handle to resize which is offset from the circle. This makes the code a bit messy, but was a necessary workaround because overlapping elements ALL fire their mouse events and it messed everything up. It is possible to check if the event has been fired, but the circle event was always firing first, when it should have been the handle event for good UX.