feat: unmount comp at end of test that check for loading state
For those tests, we only care about if loading state ever holds and not
any data that comes in afterwards that would change its state. By
unmounting at the end of the test, we ensure said data won't be changing
state anymore without being accounted for and safely fixes the act
warning.