feat: add ErrorBoundary component
This component can be used to implement error reporting to our backend service. It has a default fallback component set but might be needed to override this default for specific boundaries (such as boundary for our map?).

In the future we might add a required prop to this ErrorBoundary component to specifya 'location' of our app so errors can be grouped by their location. Eg. request form, map, edit profile, ...