test: fix typescript errors in friend request tests
Note: imho property sent should be of a enum type instead of a boolean. eg. FriendRequestKind.Sent and FriendRequestKindReceived. Although the prop might not be needed in frontend as we receive both kinds in seperate arrays.
If we do prefere boolean, isSent would be my personal preference for naming it. (prefix boolean names with 'is, should, can, may, ...)