feat: improve scroll behaviour to comment box in Safari
Even with the polyfill, Safari doesn't support the "block: center"
option of smooth scrolling, which nicely scrolls the input into centre
of the screen.

However, there's a Webkit proprietary "scrollIntoViewIfNeeded" which
when called with true, it does scroll the input to centre of screen.

I think this is a worthy tradeoff on Safari than keeping smooth
scrolling, which always smooth scroll the input to the top of screen
and get clipped off by our fixed position nav bar.