fix: useUserByUserName shouldn't throw errors when username is empty string
Small fix so that the service doesn't log errors on load of the profile page because on first render the current user is still null. In that case username was an empty string (instead of undefined to make typescript happy). I've changed the signature of useUserByUsername so it accepts also undefined as username.