chore: update our ErrorBoundary's API interface
Such that the App doesn't need to explicitly pass in a `fallbackRender`
prop. This is due to a recent type update in react-error-boundary, where
only one of `fallback`, `FallbackComponent` or `fallbackRender` can be
specified, but not more than one of them.

This is a good change since if we are not careful, providing
more than one could result in unexpected results and relies on knowing
the implementation details of the lib to determine which of the props
will win.