fix map click event handlers being called twice
When an active feature on the map was clicked, both the
`handleMapUserClick` and `handleMapClickAway` callbacks where called.

Reason for this was the order in which these event callbacks are
initiated. `handleMapUserClick` calls prevent default on the event object
and `handleMapClickAway` checks if `event.preventDefault()` was called
already. So handleMapUserClick should be bound before
handleMapClickAway.

Also removed an (imho) unneeded dependency of the useEffect hook.