fix: map feature doesn't have active styling on slow internet
Moved logic from handeResultClick callback function into a useEffect
hook so it can listen to map state changes (areClustersLoaded) and
wait for executing any map related code untill the map is loaded.

This way we can be sure that the feature of the selected user will
always be styled as active. No matter if the result was clicked before
or after the map data was loaded.

fixes #1846