feat: add edit event page test
Fixed edge case where event value is not populating form if user visits
the page directly by delaying rendering of form until the query has
loaded.

This is because the event wouldn't have been in cache yet if
user goes to the edit event form directly, and so form would use an
undefined event for first render (as if it's creating a new one).