fix: default start/end date & time not adjusting correctly at 11pm
Spotted this because the current tests are failing on develop.

Between 11-11:59pm on any given day, the existing code for making sure
the default start date is at the next hour is by adding an hour to it,
but that wraps back around to 12am and since the date part was not
adjusted, the default date and time goes back to 12am of the given day.
The end date/time suffers a similar problem (1am of the given day) since
it's relative to the start date/time.

This fixes it by checking that if we are at 11pm, the default start/end
date also gets adjusted to tomorrow.