fix: add an absolute min-width to grid column to prevent overflow
The problem seems to be a grid "gotcha", where "1fr" really means
minmax(auto, 1fr), which means it's possible that the element in the
grid's min-width is larger than the width of the grid itself. I.e. auto
is greater than 1fr. This means the grid column can't shrink below that
size.

Setting min-width to 0 with minmax(0, 1fr) at the grid level allows the
grid column to shrink to fit to the width of the container, as 1fr is
guaranteed to be the max value of the two.
https://css-tricks.com/preventing-a-grid-blowout/