Create background job type for sending event creation emails and define handler.
The handler includes some validation to ensure users are not spammed with event emails.
For example, if the event creator has not completed their profile, we abort the event creation.
It's probably a very small edge case that an email will be sent out for a created event.
I also included a filter for the largest nodes, global and regional.
I then grab all the users that are subscribed to a cluster and send an email to all of them for the event.
If there are too many users, I limit the users emailed to those in a 20km radius of the event occurrence.