revert docker port forward on host to 5000
this is more expected because it matches with the container port. i
originally did this because AirPlay runs on 5000 and i thought it made
sense to make the code work more generally for people since many have
macOS with this same issue, but it's probably better to not enforce
local differences like this on the codebase, so i just turn AirPlay off.