feat: AddAdminNote API
This implementation treats a new field on User (admin_note) as an
append-only log that prefixes plaintext with a timestamp and the id and
username of the superadmin that logged it.

This feature is intended to detail any relevant notes for moderation,
like the reason behind banning users. Because of this, we tie it to the
BanUser endpoint as well.