improve: yarn start:localdev uses docker backend
Currently the workflow for using the local backend with the local
frontend is to replace `.env.development` with `.env.localdev` which
modifies the git working tree before running `yarn start`. Naturally
this must be undone when you commit, which is generally not good
development practice and also confusing that this would be necessary if
a script that specifies the right environment config file could be used
instead.

There is an [unfortunate
shortcoming](vercel/next.js#25764) in
NextJS for environment management, which is that ["non-standard"
NODE_ENV](https://nextjs.org/docs/messages/non-standard-node-env)
variables are not supported due to possible side effects. Because of
this, and the fact that the NODE_ENV is determined from the NextJS
command that is run, you cannot actually tell `yarn dev` to use a
different environment file than `.env.development` (this must be used!).

Fortunately, there is an [environment variable load
order](https://nextjs.org/docs/pages/building-your-application/configuring/environment-variables)
which we take advantage of here to inject the `.env.localdev` variables
before the `.env.development` ones (both files are loaded, but
`localdev` ones come first).

It's a straightforward workaround, but many wish it didn't have to be
this way. I preferred this solution over installing some dependency that
isn't really necessary, especially because it follows a NextJS standard.