feat: ToolChipLink + Badge components w/ routing
I thought having a clickable chip with a tooltip is a common scenario we
might want to make use of later, so I extracted this functionality
originally built out in the features/profile/view/Badges file into a
ToolChipLink component.

The Badge component is then just a wrapper around that where you pass in
a Badge resource. This uses a URL scheme for badges at /badges/{id}